+++ src/components/cards/PoolStakesTableCard.tsx
@@ -26,7 +26,11 @@ import {
   formatCurrency,
   formatPercentage,
 } from '../../lib/utils/number';
-import { Token, getDisplayDenomAmount } from '../../lib/web3/utils/tokens';
+import {
+  Token,
+  getDisplayDenomAmount,
+  getTokenId,
+} from '../../lib/web3/utils/tokens';
 
 import { usePoolDepositFilterForPair } from '../../lib/web3/hooks/useUserShares';
 import {
@@ -37,7 +41,7 @@ import IncentivesCard from './IncentivesCard';
 
 import { tickIndexToPrice } from '../../lib/web3/utils/ticks';
 import { guessInvertedOrder } from '../../lib/web3/utils/pairs';
-import { getTokenId, matchTokens } from '../../lib/web3/hooks/useTokens';
+import { matchTokens } from '../../lib/web3/hooks/useTokens';
 import { useCurrentPriceFromTicks } from '../Liquidity/useCurrentPriceFromTicks';
 
 import './PoolsTableCard.scss';
