+++ src/pages/Pool/PoolOverview.tsx
@@ -12,7 +12,11 @@ import { SmallCardRow } from '../../components/cards/SmallCard';
 import StatCardTVL from '../../components/stats/StatCardTVL';
 
 import { formatAddress } from '../../lib/web3/utils/address';
-import { Token, getDisplayDenomAmount } from '../../lib/web3/utils/tokens';
+import {
+  Token,
+  getDisplayDenomAmount,
+  getTokenId,
+} from '../../lib/web3/utils/tokens';
 import { getPairID, hasInvertedOrder } from '../../lib/web3/utils/pairs';
 import {
   ChainEvent,
@@ -37,7 +41,6 @@ import { formatRelativeTime } from '../../lib/utils/time';
 
 import './Pool.scss';
 import useTokens, {
-  getTokenId,
   matchTokenByDenom,
   useTokenPathPart,
   useTokenValue,
