+++ src/pages/Swap/Swap.tsx
@@ -14,7 +14,6 @@ import { LimitOrderType } from '@duality-labs/dualityjs/types/codegen/dualitylab
 
 import TokenInputGroup from '../../components/TokenInputGroup';
 import {
-  getTokenId,
   getTokenPathPart,
   useTokenBySymbol,
 } from '../../lib/web3/hooks/useTokens';
@@ -33,7 +32,11 @@ import { getRouterEstimates, useRouterResult } from './hooks/useRouter';
 import { useSwap } from './hooks/useSwap';
 
 import { formatPercentage } from '../../lib/utils/number';
-import { Token, getBaseDenomAmount } from '../../lib/web3/utils/tokens';
+import {
+  Token,
+  getBaseDenomAmount,
+  getTokenId,
+} from '../../lib/web3/utils/tokens';
 import { formatLongPrice } from '../../lib/utils/number';
 
 import './Swap.scss';
